New file for usePreloadData, preload data helpers
